+++ src/ngrest/plugins/Link.php
@@ -52,7 +52,7 @@ class Link extends Plugin
      */
     public function onBeforeSave($event)
     {
-        // if its not i18n casted field we have to serialize the the file array as json and abort further event excution.
+        // if its not i18n casted field we have to serialize the file array as json and abort further event excution.
         if (!$this->i18n) {
             $this->writeAttribute($event, $this->i18nFieldEncode($event->sender->getAttribute($this->name)));
             return false;